# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "crud-item-api",
        "description": "Это пример API для управления элементами. API предоставляет методы для создания, обновления, получения и удаления элементов",
        "contact": {
            "email": "heturion@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/items": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Создать новый элемент",
                "operationId": "af37857663e6805dd367a55f374c9ee1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 15
                                    },
                                    "key": {
                                        "type": "string",
                                        "maxLength": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Элемент успешно создан"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Получить информацию о элементе",
                "operationId": "7f1ffbd9fa2b704c61ebcccb636ebb27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор элемента",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о элементе"
                    },
                    "404": {
                        "description": "Элемент не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Обновить элемент",
                "operationId": "4ac9fb55d8a18bd1cba3a29d84f9fd4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор элемента",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 15
                                    },
                                    "key": {
                                        "type": "string",
                                        "maxLength": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Элемент успешно обновлен"
                    },
                    "404": {
                        "description": "Элемент не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Удалить элемент",
                "operationId": "0f88b60b73861d5414f8ea2fedd3dfc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор элемента",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Элемент успешно удален"
                    },
                    "404": {
                        "description": "Элемент не найден"
                    }
                }
            }
        }
    }
}